feat: Factor out directory functionality into its own reusable component and adapt current pages to use it
